+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 * Add optionnal ``version`` property on package manager definitions.
 * Allow each package manager to set requirement on its own version.
 * Let ``mas`` report its own version.
+* Bump minimal requirement of ``mas`` to 1.3.1.
+* Fix parsing of ``mas`` package versions after the 1.3.1 release.
 
 
 `1.9.0 (2016-09-23) <https://github.com/kdeldycke/meta-package-manager/compare/v1.8.0...v1.9.0>`_